--- docroot/sites/registrar.uiowa.edu/modules/registrar_core/registrar_core.services.yml
@@ -1,4 +1,0 @@
-services:
-  registrar_core.lazy_builder:
-    class: Drupal\registrar_core\LazyBuilder\AcademicCalendarLazyBuilder
-    arguments: ['@renderer', '@uiowa_maui.api']
